Redesign custom order of rooms

+++ src/main/java/com/softserve/entity/Room.java
@@ -30,7 +30,6 @@ public class Room implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-
     @NotEmpty(message = "Name cannot be empty")
     @Size(min = 2, max = 35, message = "Name must be between 2 and 35 characters long")
     @Column(length = 35, nullable = false)
@@ -44,5 +43,5 @@ public class Room implements Serializable {
     private boolean disable = false;
 
     @Column(name = "sort_order")
-    private Double sortOrder;
+    private Integer sortOrder;
 }
